Rename fllter package to filter in shared/gradle domain.
Fixes issue #158: rename typo in package name from fllter to filter.
- Move BinaryInterleaver.kt and Nybbler.kt to correct package directory
- Update package declarations in both files
- Fix 5 importer statements across shared:gradle and flows domains
- All tests pass; zero fllter references remain in codebase
- Execution log and plan marked complete
