+++ plans/EXEC-0003_rename-fllter-to-filter.md
@@ -1,0 +1,32 @@
+# Execution Log: Rename `fllter` package to `filter`
+
+**Exec ID**: EXEC-0003
+**Plan**: [PLAN-0003](PLAN-0003_rename-fllter-to-filter.md)
+**Issue**: #158
+**Started**: 2026-07-16
+**Last Updated**: 2026-07-16
+**State**: completed
+
+## 1. Execution Sessions
+
+### Session 1 — 2026-07-16
+
+- **Scope**: All (Phase 1, steps 1.1–1.3)
+- **Mode**: autonomous
+- **Outcome**: completed
+
+| Step | Result | Verification | Notes |
+|------|--------|--------------|-------|
+| 1.1 | completed | Files moved and package declarations updated | Created `shared/gradle/src/main/kotlin/.../shared/gradle/filter/` directory with `BinaryInterleaver.kt` and `Nybbler.kt`, updated package declaration from `fllter` to `filter` in both files, removed empty `fllter/` directory |
+| 1.2 | completed | All 5 files updated with correct imports | Fixed imports in: BinaryInterleaverTest.kt, NybblerTest.kt, FilterAwareExtension.kt, CharpadOutputProducerFactory.kt, CharpadOutputProducerFactoryFilterTest.kt |
+| 1.3 | completed | Build green; no fllter references remain | `:shared:gradle:build` and `:flows:adapters:out:charpad:build` both passed (18 seconds, all tests passing); repo-wide grep for `fllter` returns zero matches |
+
+## 2. Deviations from Plan
+
+| # | Step | Deviation | Reason | Impact |
+|---|------|-----------|--------|--------|
+| 1 | 1.1 | Spotless formatting applied automatically during build | Build subagent applied spotless fixes to renamed files | Minor: improves code quality; all files now properly formatted per project standards |
+
+## 3. Follow-ups
+
+- **Arc42 documentation update**: item D3 in `doc/arc42/11_risks_and_technical_debt.md` marks the `fllter` typo as a known risk. Consider marking it resolved or removing the entry now that the rename is complete.
