+++ plans/PLAN-0003_rename-fllter-to-filter.md
@@ -2,7 +2,7 @@
 
 **Plan ID**: PLAN-0003
 **Issue**: #158
-**Status**: accepted
+**Status**: implemented
 **Created**: 2026-07-16
 **Last Updated**: 2026-07-16
 
@@ -178,6 +178,7 @@ None — this is a small, fully-scoped mechanical rename with no open design que
 | Date | Updated By | Changes |
 |------|------------|---------|
 | 2026-07-16 | AI Agent | Plan created and accepted; no unresolved questions, mechanical rename scope confirmed. |
+| 2026-07-16 | AI Agent | Execution completed: all 3 steps finished successfully. Package renamed, imports fixed, builds green. Status → implemented. |
 
 ---
