+++ plans/README.md
@@ -7,3 +7,4 @@ Plans are permanent artifacts — do not delete. Terminal plans (`implemented`, 
 |----|------|--------|-------|-------|------|
 | [PLAN-0001](PLAN-0001_pipeline-dsl-parallel-execution.md) | 2026-03-29 | implemented | Pipeline DSL - Enable Parallel Execution | #135 | [EXEC-0001](EXEC-0001_pipeline-dsl-parallel-execution.md) |
 | [PLAN-0002](PLAN-0002_arc42-technical-documentation.md) | 2026-07-15 | implemented | arc42 Technical Documentation (Markdown + Mermaid) | #154 | [EXEC-0002](EXEC-0002_arc42-technical-documentation.md) |
+| [PLAN-0003](PLAN-0003_rename-fllter-to-filter.md) | 2026-07-16 | implemented | Rename `fllter` package to `filter` | #158 | [EXEC-0003](EXEC-0003_rename-fllter-to-filter.md) |
