+++ shared/gradle/src/main/kotlin/com/github/c64lib/rbt/shared/gradle/dsl/FilterAwareExtension.kt
@@ -25,8 +25,8 @@ SOFTWARE.
 package com.github.c64lib.rbt.shared.gradle.dsl
 
 import com.github.c64lib.rbt.shared.domain.IllegalConfigurationException
-import com.github.c64lib.rbt.shared.gradle.fllter.BinaryInterleaver
-import com.github.c64lib.rbt.shared.gradle.fllter.Nybbler
+import com.github.c64lib.rbt.shared.gradle.filter.BinaryInterleaver
+import com.github.c64lib.rbt.shared.gradle.filter.Nybbler
 import com.github.c64lib.rbt.shared.processor.BinaryOutput
 import com.github.c64lib.rbt.shared.processor.BinaryOutputBuffer
 import com.github.c64lib.rbt.shared.processor.DevNull
