+++ shared/gradle/src/main/kotlin/com/github/c64lib/rbt/shared/gradle/filter/BinaryInterleaver.kt
@@ -22,7 +22,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-package com.github.c64lib.rbt.shared.gradle.fllter
+package com.github.c64lib.rbt.shared.gradle.filter
 
 import com.github.c64lib.rbt.shared.domain.IllegalInputException
 import com.github.c64lib.rbt.shared.processor.BinaryOutput
